+++ src/frontend/src/components/Login/LoggedIn.jsx
@@ -15,6 +15,7 @@ const useStyles = makeStyles({
     textDecoration: 'none',
     fontSize: '1.5rem',
     color: 'white',
+    lineHeight: 1,
   },
   avatar: {
     height: '1em',
